AjayP_PatternAssignments_38_to_44 Refactoring
