+++ src/ajay_PatternAssignments38to44/Assignment42.java
@@ -1,21 +1,26 @@
+/*
+Print pattern below 
+1
+2 2
+3 3 3
+4 4 4 4
+*/
 package ajay_PatternAssignments38to44;
 
-public class Pattern6 {
-	
+public class Assignment42 {
+
 	void patternPrinting() {
-		char count='A';
+
 		for(int i=1;i<=4;i++) {
-			
 			for(int j=1;j<=i;j++) {
-				System.out.print(count +" ");
-				count++;
+				System.out.print(i +" ");	
+
 			}System.out.println();
 		}
 	}
-
 	public static void main(String[] args) {
-		Pattern6 pattern6=new Pattern6();
-		pattern6.patternPrinting();
+		Assignment42 pattern5=new Assignment42();
+		pattern5.patternPrinting();
 
 	}
